chore: extract enum naming logic
