+++ gen/schema_gen_primitive.go
@@ -2,8 +2,6 @@ package gen
 
 import (
 	"fmt"
-	"strconv"
-	"unicode/utf8"
 
 	"github.com/go-faster/errors"
 
@@ -33,106 +31,16 @@ func (g *schemaGen) enum(name string, t *ir.Type, schema *jsonschema.Schema) (*i
 		return nil, errors.Wrap(err, "validate enum")
 	}
 
-	type namingStrategy int
-	const (
-		pascalName namingStrategy = iota
-		pascalSpecialName
-		cleanSuffix
-		indexSuffix
-		_lastStrategy
-	)
-
-	vstrCache := make(map[int]string, len(schema.Enum))
-	nameEnum := func(s namingStrategy, idx int, v any) (string, error) {
-		vstr, ok := vstrCache[idx]
-		if !ok {
-			vstr = fmt.Sprintf("%v", v)
-			if vstr == "" {
-				vstr = "Empty"
-			}
-			vstrCache[idx] = vstr
-		}
-
-		switch s {
-		case pascalName:
-			return pascal(name, vstr)
-		case pascalSpecialName:
-			return pascalSpecial(name, vstr)
-		case cleanSuffix:
-			return name + "_" + cleanSpecial(vstr), nil
-		case indexSuffix:
-			return name + "_" + strconv.Itoa(idx), nil
-		default:
-			panic(unreachable(s))
-		}
-	}
-
-	isException := func(start namingStrategy) bool {
-		if start == pascalName {
-			// This code is called when vstrCache is fully populated, so it's ok.
-			for _, v := range vstrCache {
-				if v == "" {
-					continue
-				}
-
-				// Do not use pascal strategy for enum values starting with special characters.
-				//
-				// This rule is created to be able to distinguish
-				// between negative and positive numbers in this case:
-				//
-				// enum:
-				//   - '1'
-				//   - '-2'
-				//   - '3'
-				//   - '-4'
-				firstRune, _ := utf8.DecodeRuneInString(v)
-				if firstRune == utf8.RuneError {
-					panic(fmt.Sprintf("invalid enum value: %q", v))
-				}
-
-				_, isFirstCharSpecial := namedChar[firstRune]
-				if isFirstCharSpecial {
-					return true
-				}
-			}
-		}
-
-		return false
-	}
-
-	chosenStrategy, err := func() (namingStrategy, error) {
-	nextStrategy:
-		for strategy := pascalName; strategy < _lastStrategy; strategy++ {
-			// Treat enum type name as duplicate to prevent collisions.
-			names := map[string]struct{}{
-				name: {},
-			}
-			for idx, v := range schema.Enum {
-				k, err := nameEnum(strategy, idx, v)
-				if err != nil {
-					continue nextStrategy
-				}
-				if _, ok := names[k]; ok {
-					continue nextStrategy
-				}
-				names[k] = struct{}{}
-			}
-			if isException(strategy) {
-				continue nextStrategy
-			}
-			return strategy, nil
-		}
-		return 0, errors.Errorf("unable to generate variant names for enum %q", name)
-	}()
+	nameGen, err := enumVariantNameGen(name, schema.Enum)
 	if err != nil {
 		return nil, errors.Wrap(err, "choose strategy")
 	}
 
 	var variants []*ir.EnumVariant
 	for idx, v := range schema.Enum {
-		variantName, err := nameEnum(chosenStrategy, idx, v)
+		variantName, err := nameGen(v, idx)
 		if err != nil {
-			return nil, errors.Wrapf(err, "variant %q [%d]", vstrCache[idx], idx)
+			return nil, errors.Wrapf(err, "variant %q [%d]", fmt.Sprintf("%v", v), idx)
 		}
 
 		variants = append(variants, &ir.EnumVariant{
